Split mixed read-only live tests
