+++ tests/Clc.Polaris.Api.LiveIntegrationTests/ReadOnly/SystemAdministration/SAMobilePhoneCarriersGetReadOnly.LiveTests.cs
@@ -1,0 +1,17 @@
+namespace Clc.Polaris.Api.LiveIntegrationTests
+{
+    [TestClass]
+    public sealed class SAMobilePhoneCarriersGetReadOnlyLiveTests : IntegrationTestBase
+    {
+        [TestMethod]
+        [ReadOnlyLiveTest]
+        public async Task SAMobilePhoneCarriersGetAsync_ReturnsResponse()
+        {
+            IntegrationTestRequirements.RequireStaffOverrideAccount(PapiSettings);
+
+            var response = await Papi.SAMobilePhoneCarriersGetAsync(cancellationToken: TestContext.CancellationToken);
+
+            Assert.IsNotNull(response.Data);
+        }
+    }
+}
