+++ tests/Clc.Polaris.Api.LiveIntegrationTests/ReadOnly/SystemAdministration/SysHoldStatusesGetReadOnly.LiveTests.cs
@@ -1,0 +1,15 @@
+namespace Clc.Polaris.Api.LiveIntegrationTests
+{
+    [TestClass]
+    public sealed class SysHoldStatusesGetReadOnlyLiveTests : IntegrationTestBase
+    {
+        [TestMethod]
+        [ReadOnlyLiveTest]
+        public async Task SysHoldStatusesGetAsync_ReturnsResponse()
+        {
+            var response = await Papi.SysHoldStatusesGetAsync(cancellationToken: TestContext.CancellationToken);
+
+            Assert.IsNotNull(response.Data);
+        }
+    }
+}
